src/requirements.txt
minio==7.1.12
pandas==1.4.3
python-dotenv==0.21.0
scipy==1.7.3
fastapi==0.75.2
numpy==1.20.3
uvicorn==0.17.6
scikit-learn == 1.1.2
fastapi-utils==0.2.1
requests==2.27.1